+++ src/main/java/org/openlmis/referencedata/util/messagekeys/MessageKeys.java
@@ -58,6 +58,7 @@ public abstract class MessageKeys {
   protected static final String MUST_BE_UNIQUE = "mustBeUnique";
   protected static final String MUST_BE_PROVIDED = "mustBeProvided";
   protected static final String MISMATCH = "mismatch";
+  protected static final String INVALID_LENGTH = "invalidLength";
 
   // Entities
   protected static final String LOT = "lot";